pyproject.toml
[project]
name = "ai-issue"
version = "0.1.0"
description = "GitHub Action для автоматического создания issue на основе PR при помощи AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygithub>=2.5.0",
    "openai>=1.58.0",
    "pydantic>=2.10.0",
    "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.2",
    "types-requests>=2.32.0.20241229",
]


[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
namespace_packages = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = "github.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_issue"]